requirements.txt
requests>=2.32.4,<3.0.0
numpy>=2.3.0,<3.0.0
pandas>=2.3.0,<3.0.0
langchain-community>=0.3.24,<0.4.0
faiss-cpu>=1.11.0,<2.0.0
langchain-openai>=0.3.0,<0.4.0
python-dotenv>=1.1.0,<2.0.0
sentence-transformers>=2.6.1,<3.0.0
langchain-text-splitters>=0.3.8,<0.4.0
langchain>=0.3.25,<0.4.0
streamlit>=1.47.0,<2.0.0
uvicorn>=0.34.3,<0.35.0
